requirements.txt
mysql-connector-python
paramiko
python-dotenv
